display MCQ as dialogue: major improvements + UnreadCount minor fix
* two step validation: selectChoice, then validateReply
* show typing.gif for few second to delay result display
* add validated boolean field to Reply
* with the help of this brand new boolean, refactor & simplify "hasBeenSelected"-like ChoiceDescriptor condition
* mcq-view, whview and tabview: add ATTRS to customize validate button text
* refactor UnreadCount to have several counter on the same widget

+++ wegas-core/src/main/java/com/wegas/mcq/ejb/QuestionDescriptorFacade.java
@@ -295,8 +295,12 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
     }
 
     private Reply internalCancelReply(Reply reply) {
-        this.getEntityManager().remove(reply);
-        return reply;
+        if (!reply.isValidated()) {
+            this.getEntityManager().remove(reply);
+            return reply;
+        } else {
+            throw WegasErrorMessage.error("Cannot cancel a validated reply");
+        }
     }
 
     /**
@@ -327,16 +331,21 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
     }
 
     /**
-     * create a reply for given player based on given choice
+     * create a reply for given player based on given choice.
+     * <p>
+     * Pre existing not validated reply will be cancelled if the question accept one reply only.
      *
      * @param choiceId  selected choice
      * @param player    player who select the choice
      * @param startTime time the player select the choice
      *
      * @return the new reply
+     *
+     * @throws WegasErrorMessage if the question is fully validated or the maximum number of replies has been reached
      */
     @Override
-    public Reply selectChoice(Long choiceId, Player player, Long startTime) {
+    public Reply selectChoice(Long choiceId, Player player, Long startTime)
+            throws WegasErrorMessage {
         ChoiceDescriptor choice = getEntityManager().find(ChoiceDescriptor.class, choiceId);
         QuestionDescriptor questionDescriptor = choice.getQuestion();
         QuestionInstance questionInstance = questionDescriptor.getInstance(player);
@@ -364,9 +373,7 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
                 toCancel.add(r);
             }
 
-            /*
-             * Two steps deletion avoids concurrent modification exception
-             */
+            /* Two steps deletion avoids concurrent modification exception */
             for (Reply r : toCancel) {
                 this.cancelReply(player.getId(), r.getId());
             }
@@ -389,6 +396,7 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
         }
 
         Reply reply = this.createReply(choiceId, player, startTime, false);
+
         try {
             scriptEvent.fire(player, "replySelect", new ReplyValidate(reply));
         } catch (WegasScriptException e) {
@@ -423,6 +431,42 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
     @Override
     public Reply selectChoice(Long choiceId, Long playerId, Long startTime) {
         return this.selectChoice(choiceId, playerFacade.find(playerId), startTime);
+    }
+
+    /**
+     * {@link #selectChoice(java.lang.Long, java.lang.Long, java.lang.Long) selectChoice}
+     * with startTime = 0
+     *
+     * @param choiceId
+     * @param player
+     * @param startTime
+     *
+     * @return the new reply
+     */
+    @Override
+    public Reply deselectOthersAndSelectChoice(Long choiceId, Player player, Long startTime) {
+        ChoiceDescriptor choice = getEntityManager().find(ChoiceDescriptor.class, choiceId);
+        QuestionDescriptor questionDescriptor = choice.getQuestion();
+        QuestionInstance questionInstance = questionDescriptor.getInstance(player);
+
+        List<Reply> pendingReplies = questionInstance.getReplies(player, Boolean.FALSE);
+        for (Reply r : pendingReplies) {
+            this.cancelReplyTransactional(player, r);
+        }
+
+        return this.selectChoice(choiceId, player, startTime);
+    }
+
+    /**
+     * @param choiceId  selected choice id
+     * @param playerId  id of player who select the choice
+     * @param startTime
+     *
+     * @return the new reply
+     */
+    @Override
+    public Reply deselectOthersAndSelectChoice(Long choiceId, Long playerId, Long startTime) {
+        return this.deselectOthersAndSelectChoice(choiceId, playerFacade.find(playerId), startTime);
     }
 
     /**
@@ -515,37 +559,44 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
 
     /**
      * @param player
-     * @param validateReply
+     * @param reply
      *
-     * @throws com.wegas.core.exception.client.WegasRuntimeException
+     * @throws WegasRuntimeException
      */
     @Override
-    public void validateReply(final Player player, final Reply validateReply) throws WegasRuntimeException {
-        final ChoiceDescriptor choiceDescriptor = validateReply.getResult().getChoiceDescriptor();
-        validateReply.setResult(choiceDescriptor.getInstance(player).getResult());// Refresh the current result
+    public Reply validateReply(final Player player, final Reply reply) throws WegasRuntimeException {
+        if (!reply.isValidated()) {
+            reply.setValidated(true);
 
-        if (validateReply.getIgnored()) {
-            scriptManager.eval(player, validateReply.getResult().getIgnorationImpact(), choiceDescriptor);
-        } else {
-            scriptManager.eval(player, validateReply.getResult().getImpact(), choiceDescriptor);
-        }
-        ChoiceInstance choiceInstance = validateReply.getChoiceInstance();
+            final ChoiceDescriptor choiceDescriptor = reply.getResult().getChoiceDescriptor();
+            reply.setResult(choiceDescriptor.getInstance(player).getResult());// Refresh the current result
 
-        final ReplyValidate replyV = new ReplyValidate(validateReply, choiceInstance,
-                (QuestionInstance) choiceDescriptor.getQuestion().getInstance(player),
-                player);
-        try {
-            requestManager.addUpdatedEntities(choiceDescriptor.getEntities());
-            if (validateReply.getIgnored()) {
-                scriptEvent.fire(player, "replyIgnore", replyV);
+            if (reply.getIgnored()) {
+                scriptManager.eval(player, reply.getResult().getIgnorationImpact(), choiceDescriptor);
             } else {
-                scriptEvent.fire(player, "replyValidate", replyV);
+                scriptManager.eval(player, reply.getResult().getImpact(), choiceDescriptor);
             }
-        } catch (WegasRuntimeException e) {
-            logger.error("EventListener error (\"replyValidate\")", e);
-            // GOTCHA no eventManager is instantiated
+            ChoiceInstance choiceInstance = reply.getChoiceInstance();
+
+            final ReplyValidate replyV = new ReplyValidate(reply, choiceInstance,
+                    (QuestionInstance) choiceDescriptor.getQuestion().getInstance(player),
+                    player);
+            try {
+                requestManager.addUpdatedEntities(choiceDescriptor.getEntities());
+                if (reply.getIgnored()) {
+                    scriptEvent.fire(player, "replyIgnore", replyV);
+                } else {
+                    scriptEvent.fire(player, "replyValidate", replyV);
+                }
+            } catch (WegasRuntimeException e) {
+                logger.error("EventListener error (\"replyValidate\")", e);
+                // GOTCHA no eventManager is instantiated
+            }
+            this.replyValidate.fire(replyV);
+            return reply;
+        } else {
+            throw WegasErrorMessage.error("This reply has already been validated");
         }
-        this.replyValidate.fire(replyV);
     }
 
     /**
@@ -553,8 +604,8 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
      * @param replyVariableInstanceId
      */
     @Override
-    public void validateReply(Player player, Long replyVariableInstanceId) {
-        this.validateReply(player, getEntityManager().find(Reply.class, replyVariableInstanceId));
+    public Reply validateReply(Player player, Long replyVariableInstanceId) {
+        return this.validateReply(player, getEntityManager().find(Reply.class, replyVariableInstanceId));
     }
 
     /**
@@ -562,8 +613,8 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
      * @param replyVariableInstanceId
      */
     @Override
-    public void validateReply(Long playerId, Long replyVariableInstanceId) {
-        this.validateReply(playerFacade.find(playerId), replyVariableInstanceId);
+    public Reply validateReply(Long playerId, Long replyVariableInstanceId) {
+        return this.validateReply(playerFacade.find(playerId), replyVariableInstanceId);
     }
 
     public void validateQuestion(final VariableInstance question, final Player player) throws WegasRuntimeException {
@@ -673,6 +724,7 @@ public class QuestionDescriptorFacade extends BaseFacade<ChoiceDescriptor> imple
      * @param questionInstanceId
      * @param playerId
      */
+    @Override
     public void validateQuestion(Long questionInstanceId, Long playerId) {
         this.validateQuestion(questionInstanceId, playerFacade.find(playerId));
     }
